put navbar and topbar into seperate files, changed styling in accordance with figma
